Fixed the issue of netcdf_fileout. Updated version number to 3.10.1

+++ HDFCFUtil.cc
@@ -306,7 +306,7 @@ HDFCFUtil::print_type(int32 type)
         return BYTE;
 
     case DFNT_INT8:
-        return INT16;
+        return INT32;
 
     case DFNT_UINT16:
         return UINT16;
